Replaced .json() calls .json property gets.
We'd been using this script for a long time (thanks!) and last year it started failing.  The error was, for example...

Traceback (most recent call last):
  File "trello-backup.py", line 127, in <module>
    main()
  File "trello-backup.py", line 81, in main
    if len(boards.json()) <= 0:
TypeError: 'list' object is not callable

It seems that Request.json is actually an object property, not a method.  After making the modifications seen in this commit, the script started working again.